# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LaunchMailer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name, email, phone, city, or tags
        - name: hasEmail
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
        - name: hasPhone
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of tags to filter by
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
  /customers/import:
    post:
      operationId: importCustomers
      tags: [customers]
      summary: Import customers from parsed CSV/Excel data
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /customers/export:
    get:
      operationId: exportCustomers
      tags: [customers]
      summary: Export customers as CSV
      security:
        - bearerAuth: []
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a single customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CAMPAIGNS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        unsubscribed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string

        phone:
          type: string
        city:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    CustomerImportInput:
      type: object
      required: [customers, updateExisting]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/CustomerRow"
        updateExisting:
          type: boolean
          description: Whether to update existing customers on email/phone match
    CustomerRow:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        tags:
          type: string
          description: Comma-separated list of tags
        notes:
          type: string
    CustomerListResponse:
      type: object
      required: [customers, total, page, limit]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ImportResult:
      type: object
      required: [created, updated, skipped, errors, uploadId]
      properties:
        created:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
        uploadId:
          type: string
    ImportError:
      type: object
      required: [row, field, message]
      properties:
        row:
          type: integer
        field:
          type: string
        message:
          type: string
        value:
          type: ["string", "null"]

    # ─── CAMPAIGNS ────────────────────────────────────────────────────────────
...